vim: when hiding buffer, delete if empty
When `set hidden` is on and you run `vim --remote file1` it creates an
initial blank buffer before opening file1. This means you start with 2
buffers instead of just the one you opened. This checks and if the
buffer is blank it will delete it. It shouldn't delete a blank buffer
that has been modified.