vim: Convert submodules to local files
This removes the submodules I had setup for vim bundles and instead
contains the actual code. This solves my concern about a submodule going
away and no longer having access to it. Which also means I no longer
need to maintain a fork of it for the same reason.

Note this method removed all the .git* files in submodules, including
.gitignore. So may need to update the repo's .gitignore.