firefox: Fix user prefs and reset network tweaks
First the trimURLs option to be false which is what I intended.

Commented out all the various network prefs I set 8 or 9 years ago since
nowadays everyone is on a fast connection so those should no longer be
needed.  Remember to go in to about:config and manually reset all those
settings