firefox: Extend timeout on javascript
When starting up firefox I'll get a popup about a script taking too
long and if I should continue or abort. I think it's a bug in lastpass
addon since that's always the script that's taking so long. I have like
212 tabs at the moment (treestyletabs ftw) and I think lastpass is
scanning each and every one of those for a login field, at startup.
Opened ticket but they don't seem to think it's them... at any rate it
likely only comes up for people with 100s of tabs like me. This extends
the timeout to 20 seconds instead of 10 and that has seemed to fix the
popup message