git: Fix bug in `sweep` alias
It was excluding if the names of common branches were ANYWHERE in the
name.  For example a branch called `fix_broken_master_merge` would never
get cleaned up.  Fixed so now it matches the full line.

There's still an issue that if the name begins with one of the selected
names it won't match.  This is good for matching both `develop` and
`development` but bad for `master_updates`.