firefox: Update autoplay settings and a couple others
In version 63 of firefox they redid the media autoplay stuff.  This
sets it to be in line with how it worked before.  The
`media.autoplay.enabled.user-gestures-needed` being true causes
some bugs but also makes it more usable since it was harder to play
videos if that was off.

Also remove loop option (was removed  while back) and noticed a
couple more settings that were added recently.