Fix list_installed and python checks to work with pyenv
With pyenv installed running a python command like `python3` will point
to a real file (so command check passes) but it's just telling you which
environment that command is in.  Instead try to run the command as it
will fail if it's just saying it's not in a given environment.